+++ src/color.rs
@@ -24,6 +24,14 @@ pub fn cyan_string(str: &str) -> String {
     String::from(str).bright_cyan().to_string()
 }
 
+pub fn purple_string(str: &str) -> String {
+    if cfg!(test) {
+        return normal_string(str);
+    }
+
+    String::from(str).purple().to_string()
+}
+
 pub fn blue_string(str: &str) -> String {
     if cfg!(test) {
         return normal_string(str);